actually move SD dependency to a separate file
it is necessary to put everything in the header so it doesn't get compiled into the library.
