+++ codecs.h
@@ -46,7 +46,6 @@
 #include <Arduino.h>
 #include <AudioStream.h>
 #include <spi_interrupt.h>
-#include <SD.h>
 
 #define ERR_CODEC_NONE				0
 #define ERR_CODEC_FILE_NOT_FOUND    1
@@ -91,47 +90,6 @@ public:
 	virtual size_t fposition(void);
 	virtual size_t fsize(void);
 	virtual size_t fread(uint8_t buffer[],size_t bytes);
-};
-
-class CodecFile : public CodecFileBase
-{
-public:
-
-	bool fopen(const char *filename) {ftype=codec_file; AudioStartUsingSPI(); fptr=NULL; file=SD.open(filename); _fsize=file.size(); _fposition=0; return file != 0;} //FILE
-	bool fopen(const uint8_t*p, const size_t size) {ftype=codec_flash; fptr=(uint8_t*)p; _fsize=size; _fposition=0; return true;} //FLASH
-	bool fopen(const size_t p, const size_t size) {ftype=codec_serflash; offset=p; _fsize=size; _fposition=0; AudioStartUsingSPI(); serflashinit(); return true;} //SERIAL FLASH
-	void fclose(void)
-	{
-		_fsize=_fposition=0; fptr=NULL;
-		if (ftype==codec_file) {file.close(); AudioStopUsingSPI();}
-		else
-		if (ftype==codec_serflash) {AudioStopUsingSPI();}
-		ftype=codec_none;
-	}
-	bool f_eof(void) {return _fposition >= _fsize;}
-	bool fseek(const size_t position) {_fposition=position;if (ftype==codec_file) return file.seek(_fposition)!=0; else return _fposition <= _fsize;}
-	size_t fposition(void) {return _fposition;}
-	size_t fsize(void) {return _fsize;}
-	size_t fread(uint8_t buffer[],size_t bytes);
-
-protected:
-//private:
-
-	void serflashinit(void);
-	void readserflash(uint8_t* buffer, const size_t position, const size_t bytes);
-
-	SPISettings spisettings;
-
-	codec_filetype ftype;
-
-	File file;
-	union {
-		uint8_t* fptr;
-		size_t offset;
-	};
-
-	size_t _fsize;
-	size_t _fposition;
 };
 
 class AudioCodec : public AudioStream
